fix(background): Chrome用Offscreenドキュメント機能とFirefox用分岐を追加
- offscreen-chrome.tsにChrome用Offscreenドキュメントの生成とsendToOffscreenを実装
- offscreen-client.tsで__BROWSER_TARGET__に応じてFirefox向けエラー分岐を追加
